backend/requirements.txt
# requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose
python-socketio[asyncio]
aiofiles
